+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Other
 
+- working on calibre.... - (Release 0.1.35) (pylnlib_commit=d1fe153)
+- working on calibre.... - (Release 0.1.35) (pylnlib_commit=5c0bb2f)
+
+## [0.1.35] - 2026-08-08
+
+## Other
+
 - working on calibre.... - (Release 0.1.35) (pylnlib_commit=5c0bb2f)
 
 ## [0.1.35] - 2026-08-07